+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Support Minecraft 1.21.6+
+- Temporary fix for sign crash issue (Fixes #34)